feat(champions.ts): add list of champions to improve data handling in the application
The `champions.ts` file was added to the `src/lib/utils/data-lol` directory. This file contains an array of champion names used in the application. The addition of this file improves the data handling in the application by providing a centralized list of champions that can be easily accessed and used throughout the codebase.