feat(attempt): add Attempt component
This commit adds a new component called Attempt. The Attempt component is responsible for rendering an attempt in a game. It receives a championName prop, which represents the name of the champion being attempted.

The Attempt component imports the necessary dependencies, including Image from "next/image" and Badge from "@/lib/components/ui/badge". It also imports types from "./attempt.type" and ReactElement from "react". Additionally, it imports functions and data from "@/lib/utils/functions/formatName", "@/lib/utils/functions/getChampionsAssets", and "@/lib/utils/data-lol/champions".

The Attempt component renders a list item with a flex layout and justify-between alignment. Inside the list item, there is a div with a flex layout and items-center alignment. It contains an Image component that displays the square assets of a champion from League of Legends. The Image component receives the champion's image URL, width, height, and alt text. The championName is also displayed next to the image. Finally, there is a Badge component that displays either "WIN" or "FAIL" based on whether the attempt was successful or not. The variant of the Badge component is determined by the isWin variable, which compares the formattedChampionName with the answerBlurredChampion from the AnswerBlurChampionStore.

The Attempt component does not have a newline at the end of the file.

This commit also includes the addition of the AttemptProps type in the attempt.type.ts file. The AttemptProps type defines the championName prop as a string.

Lastly, the index.ts file is added to export the Attempt component.