feat(gamesList.ts): add gamesList data file to store information about available games
The `gamesList.ts` file is added to the `src/lib/utils/data` directory. It defines a `GameList` type which represents the structure of each game in the list. The `gamesList` array contains information about available games, including their visual name, path, and status. Three games are initially included in the list: "Blurry Champions" with a path of "blurry-champions" and status "available", "Blurry Spells" with a path of "blurry-spells" and status "wip", and "Another One" with a path of "another-one" and status "unavailable".