chore(package.json): add "@radix-ui/react-alert-dialog" dependency feat(page.tsx): add support for displaying an alert dialog with information about the colors feat(alert-dialog.tsx): add implementation of the AlertDialog component and its related components
The changes were made to enhance the functionality of the application. The addition of the "@radix-ui/react-alert-dialog" dependency allows for the display of an alert dialog. The alert dialog is then implemented in the "page.tsx" file to show information about the colors. The "alert-dialog.tsx" file contains the implementation of the AlertDialog component and its related components. This allows for the creation and customization of alert dialogs throughout the application.