chore(package.json): add @radix-ui/react-label package to dependencies
feat(reportProposal): add ReportProposal component to allow users to send bug reports or feature proposals

refactor(page.tsx): comment out Statistics component and add ReportProposal component to the page

feat(dialogReportProposal): add Label and Textarea components to the dialog for users to input their message

feat(ui/label): create Label component to style form labels

feat(ui/textarea): create Textarea component to style textareas

The changes were made to improve the user experience by adding the ability for users to send bug reports or feature proposals through the ReportProposal component. The Label and Textarea components were added to the DialogReportProposal component to enhance the styling of the form. The Statistics component was commented out in the page.tsx file to temporarily remove it from the page. The @radix-ui/react-label package was added to the dependencies in the package.json file to support the Label component.