chore(layout.tsx): refactor metadata import to use a separate metadata.config file for better organization and reusability feat(metadata.config.ts): create metadata.config file to define metadata for the LOL-GAME website
The changes in `layout.tsx` refactor the import of the `metadata` object to use a separate `metadata.config` file located in the `config/metadata` directory. This change improves code organization and makes it easier to reuse the `metadata` object in other parts of the application.

The new `metadata.config.ts` file defines the `metadata` object, which contains information about the LOL-GAME website such as the title, description, site name, theme color, and Open Graph and Twitter metadata. This file allows for easy customization and centralization of the metadata configuration.

These changes were made to improve code organization and reusability of the metadata configuration in the LOL-GAME website.